feat(business-data-fields): migrate Fields entity to slice arch with unit and integration tests
