+++ apps/business-data-fields/src/Fields/CreateField/events/FieldCreatedEvent.ts
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) 2022, Turnly (https://turnly.app)
+ * All rights reserved.
+ *
+ * Licensed under BSD 3-Clause License. See LICENSE for terms.
+ */
+import { EntityAttributes, Event, EventType } from '@turnly/shared'
+import { Field } from 'Fields/Shared/domain/entities/Field'
+
+type Payload = EntityAttributes<Field>
+
+export class FieldCreatedEvent extends Event<Payload> {
+  public constructor(payload: Payload) {
+    super(EventType.CREATE, payload)
+  }
+}
