+++ apps/business-data-fields/src/Fields/SearchCustomerFieldsByService/api/SearchCustomerFieldsByServiceValidator.ts
@@ -1,0 +1,12 @@
+/**
+ * Copyright (c) 2022, Turnly (https://turnly.app)
+ * All rights reserved.
+ *
+ * Licensed under BSD 3-Clause License. See LICENSE for terms.
+ */
+import { Validator } from '@turnly/shared'
+
+export const SearchCustomerFieldsByServiceValidator = Validator.object({
+  serviceId: Validator.isId(),
+  organizationId: Validator.isId(),
+})
