+++ apps/business-data-fields/src/Fields/Shared/domain/contracts/IFieldsMapper.ts
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) 2022, Turnly (https://turnly.app)
+ * All rights reserved.
+ *
+ * Licensed under BSD 3-Clause License. See LICENSE for terms.
+ */
+import { IEntityMapper } from '@turnly/shared'
+import { Field } from 'Fields/Shared/domain/entities/Field'
+
+export type IFieldsMapper<Model> = IEntityMapper<Field, Model>
