+++ apps/business-data-fields/src/Fields/Shared/domain/enums/FieldTypes.ts
