+++ apps/business-data-fields/src/Fields/Shared/infrastructure/dependency/register-dependencies.ts
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) 2022, Turnly (https://turnly.app)
+ * All rights reserved.
+ *
+ * Licensed under BSD 3-Clause License. See LICENSE for terms.
+ */
+
+/**
+ * Handlers
+ *
+ * @description This file is used to register handlers to the bus.
+ */
+import { commandBus, eventBus, queryBus } from '@turnly/shared'
+
+import { FieldsModule } from '../../../FieldsModule'
+
+/**
+ * Field module
+ */
+queryBus.register(FieldsModule.getQueryHandlers())
+commandBus.register(FieldsModule.getCommandHandlers())
+eventBus.subscribe(FieldsModule.getEventSubscribers())
