--- apps/business-data-fields/src/Fields/infrastructure/api/rpc/index.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) 2022, Turnly (https://turnly.app)
- * All rights reserved.
- *
- * Licensed under BSD 3-Clause License. See LICENSE for terms.
- */
-export * from './FieldsMapper'
-export * from './FieldsServer'
