--- apps/business-data-fields/src/Fields/infrastructure/register-dependencies/dependencies.ts
@@ -1,29 +1,0 @@
-/**
- * Copyright (c) 2022, Turnly (https://turnly.app)
- * All rights reserved.
- *
- * Licensed under BSD 3-Clause License. See LICENSE for terms.
- */
-import { Box, ioc } from '@turnly/shared'
-
-import { SearchCustomerFieldsByServiceQueryHandler } from '../../application/queries/SearchCustomerFieldsByServiceQuery'
-import { FieldsController } from '../api/controllers/FieldsController'
-import { FieldsMapper } from '../persistence/mongo/entity-model-mappers/FieldMapper'
-import { FieldsReadableRepo } from '../persistence/mongo/repositories/FieldsReadableRepo'
-import { FieldsWritableRepo } from '../persistence/mongo/repositories/FieldsWritableRepo'
-
-Box.register({
-  fieldsMapper: ioc.asClass(FieldsMapper).singleton(),
-  fieldsReadableRepo: ioc.asClass(FieldsReadableRepo).singleton(),
-  fieldsWritableRepo: ioc.asClass(FieldsWritableRepo).singleton(),
-  fieldsController: ioc.asClass(FieldsController).singleton(),
-})
-
-/**
- * Query handlers
- */
-Box.register({
-  findCustomerFieldsByServiceQueryHandler: ioc
-    .asClass(SearchCustomerFieldsByServiceQueryHandler)
-    .singleton(),
-})
